warning cleanup

+++ io42io3/games/kca.h
@@ -7,12 +7,12 @@
 #include "io42io3/config.h"
 
 struct kca_config {
-    int steering_speed;
-    int steering_adc;
-    int steering_invert;
-    int steering_deadzone;
-    int button_telegraph_up;
-    int button_telegraph_down;
+    uint32_t steering_speed;
+    uint8_t steering_adc;
+    uint8_t steering_invert;
+    uint32_t steering_deadzone;
+    uint8_t button_telegraph_up;
+    uint8_t button_telegraph_down;
 };
 
 void game_kca_config_load(
